test works

+++ Infastructure/DataModels/CommentModel.cs
@@ -5,7 +5,7 @@ namespace Infastructure;
 
 public class Comment
 {
-    public BigInteger id { get; set; }
+    public int id { get; set; }
     public int post_id { get; set; }
     public int author_id { get; set; }
     [MinLength(3)]
